yolo-service/requirements.txt
ultralytics>=8.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
